Remove pk if IntegriyError is raised
If the pk is still set, then the declarative_content.content.q() will
attempt to use the pk to find the duplicate object. Since the object was
not saved to the db, this will raise an IntegrityError that is not
handled. By setting pk to None, q() will instead query by the natural
key, which will find the dupe.

[noissue]